requirements.txt
ace-tools==0.0
asttokens==3.0.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
db-dtypes==1.3.1
debugpy==1.8.9
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
flask==3.0.3
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
isodate==0.7.2
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas-gbq==0.25.0
parso==0.8.4
platformdirs==4.3.6
prompt-toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
pure-eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydata-google-auth==1.9.0
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
# pywin32==308
pyzmq==26.2.0
rdflib==7.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
beautifulsoup4==4.12.3
